pyproject.toml
[project]
name = "midaas"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jonhnatha Trigueiro", email = "jon@preludian.com.br"},
]
dependencies = [
    "pynamodb>=6.0.1",
    "fastapi[standard]>=0.115.0",
    "boto3>=1.35.37",
    "pydantic-settings>=2.6.1",
    "asaaspy>=0.28.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "Proprietary"}


[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-coverage>=0.0",
    "pytest-cov>=6.0.0",
]
[tool.pdm]
distribution = false
